Finish refactoring code based on high-level design
The code is runnable once again. The input prompts have been
changed to work for any piece. There has also been a lot of
low-level refactoring done to the code for moving pawns, which
has been moved into the move_pawn function and now works with
the new input prompts -- nearly all the old
code for the pawn error checks has been rewritten.

Also, it is now possible to move all of the pieces, but there are
no error checks for the pieces other than the pawns, so they can be
moved anywhere on the board.