Add code to allow user to abort from moving a piece
It is possible to pick a piece to move that has no valid moves.
In this case, the user can now type "abort" or "new piece" to
choose a new piece to move. In the previous version of the code,
the user would be stuck in an infinite loop if they chose a piece
with no valid moves.