Add partial checkmate detection
At the beginning of a player's turn, if the player is in check,
the program checks all squares surrounding the king to see whether
the king has a safe square to move to. However, the program does
not check whether the attacking piece can be counter-attacked or
blocked. Shravan and I have decided that we will likely not code
this during the course of our tutoring sessions.

I also rearranged the code for printing messages at the beginning
of a player's turn.